refactor: io_ops to nested modules with 'where' parsers in submodules
Keep 'what' parsers (push/pop/load/store) in parent module with general unit tests and move 'where' parsers (adv/constant/env/local/mem) to submodules along with their unit tests.
